중복 제거 및 공통 로직 추상화
