+++ template-method-pattern/coffee.ts
@@ -1,5 +1,4 @@
 import { CaffeinBeverage } from "./caffein.beverage";
-import * as readline from "readline";
 
 export class Coffee extends CaffeinBeverage {
     brew(): void {
@@ -10,23 +9,5 @@ export class Coffee extends CaffeinBeverage {
         console.log("설탕과 우유를 추가하는 중");
     }
 
-    async customerWantsCondiments(): Promise<boolean> {
-        const answer = await this.getUserInput();
-
-        return answer.toLowerCase() === "y";
-    }
-
-    private async getUserInput(): Promise<string> {
-        const rl = readline.createInterface({
-            input: process.stdin,
-            output: process.stdout,
-        });
-
-        return new Promise((resolve) => {
-            rl.question("커피에 우유와 설탕을 넣을까요? (y/n): ", (answer: string) => {
-                rl.close();
-                resolve(answer);
-            });
-        });
-    }
+    condiments = "우유와 설탕";
 }
